Create README.md
The Technology
It is an advanced artificial Intelligence system prototype, controlled via voice commands. With the help of voice commands vision is controlled, audio feedback is given to the voice commands fed as input and the program communicates with the micro-controller to automate devices. In this project four programs were running at the same time. The speech recognition accuracy of virtual assistance is too fast because it has been programmed to an integrated level. Machine learning algorithms were implemented to increase the accuracy.